move new util functions into blpapi_utils.cpp
